+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Truly fixed ability to create roads from spots that don't border. Previously it was fixed
   only on the entity level, while what really matters is interactor level. 
 - Updated score calculator - city grants two victory points
+- City gives its owner 2 of resources 
 
 ## [0.10.0]
 ### Added